Moved app contract execute handlers
