+++ contracts/app/andromeda-app-contract/src/lib.rs
@@ -6,3 +6,5 @@ pub mod mock;
 
 #[cfg(test)]
 pub mod testing;
+
+mod execute;
